+++ groovyConsole/history.md
@@ -4,6 +4,7 @@
 
 * Changed 'freeplaneVersionFrom' to a lower number, because I realized it worked in v1.9 versions too
 * Added a command to call groovyConsole without restrictions
+* Added button to reload script from disk.
 
 ## v0.0.4